修正停止扫描要等当前目录扫描完成问题；删除不存在文件报错

+++ DuplicationFilesManager/FileItem.cs
@@ -58,7 +58,16 @@ namespace DuplicationFilesManager
 
         public void Delete()
         {
-            Microsoft.VisualBasic.FileIO.FileSystem.DeleteFile(fileSetting.FullName, Microsoft.VisualBasic.FileIO.UIOption.OnlyErrorDialogs, Microsoft.VisualBasic.FileIO.RecycleOption.SendToRecycleBin);
+            try
+            {
+                if (fileSetting.Exists)
+                    Microsoft.VisualBasic.FileIO.FileSystem.DeleteFile(fileSetting.FullName, Microsoft.VisualBasic.FileIO.UIOption.OnlyErrorDialogs, Microsoft.VisualBasic.FileIO.RecycleOption.SendToRecycleBin);
+
+            }
+            catch
+            {
+
+            }
         }
 
         /// <summary>
